Refactor collector/mod.rs into multiple files, improving clarity
Prereq for removing Arc<GcData>
(if we manage that memory manually, we can make AtomicGc faster)
