+++ src/collector/mod.rs
@@ -1,28 +1,27 @@
 mod alloc;
-mod atomic_protection;
+mod collect_impl;
+mod data;
 mod dropper;
 mod trigger;
 
-use std::cmp;
-use std::hash::{Hash, Hasher};
+use std::hash::Hash;
 use std::sync::atomic::{AtomicBool, AtomicPtr, AtomicU64, Ordering};
 use std::sync::Arc;
 use std::thread::spawn;
 
-use crossbeam::queue::SegQueue;
 use crossbeam::Sender;
 use dashmap::DashMap;
-use dynqueue::DynQueue;
 use once_cell::sync::Lazy;
-use parking_lot::{Mutex, MutexGuard};
-use rayon::iter::{IntoParallelIterator, ParallelBridge, ParallelIterator};
+use parking_lot::Mutex;
 
 use crate::collector::alloc::GcAllocation;
-pub use crate::collector::atomic_protection::*;
 use crate::collector::dropper::{BackgroundDropper, DropMessage};
 use crate::collector::trigger::GcTrigger;
-use crate::lockout::{ExclusiveWarrant, Lockout, LockoutProvider, Warrant};
+use crate::concurrency::atomic_protection::{APSInclusiveGuard, AtomicProtectingSpinlock};
+use crate::concurrency::lockout::{ExclusiveWarrant, Lockout, Warrant};
 use crate::{Finalize, Scan};
+
+pub use crate::collector::data::{GcData, GcHandle, UnderlyingData};
 
 /// Intermediate struct. `Gc<T>` holds a `InternalGcRef`, which references a `GcHandle`
 /// There should be one `GcHandle` per `Gc<T>`
@@ -86,116 +85,6 @@ struct TrackedData {
     data: DashMap<Arc<GcData>, ()>,
     /// a set storing metadata on each live handle (`Gc<T>`) the collector is managing
     handles: DashMap<Arc<GcHandle>, ()>,
-}
-
-/// Represents a piece of data tracked by the collector
-#[derive(Debug)]
-pub struct GcData {
-    unique_id: u64,
-    /// a wrapper to manage (ie deallocate) the underlying allocation
-    underlying_allocation: GcAllocation,
-    /// lockout to prevent scanning the underlying data while it may be changing
-    lockout: Lockout,
-    /// have we started deallocating this piece of data yet?
-    deallocated: AtomicBool,
-    // During what collection was this last marked?
-    //     0 if this is a new piece of data
-    last_marked: AtomicU64,
-}
-
-impl GcData {
-    pub(crate) fn scan_ptr(&self) -> *const dyn Scan {
-        self.underlying_allocation.scan_ptr
-    }
-}
-
-impl LockoutProvider for Arc<GcData> {
-    fn provide(&self) -> &Lockout {
-        &self.lockout
-    }
-}
-
-impl Hash for GcData {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.unique_id.hash(state);
-    }
-}
-
-impl PartialEq for GcData {
-    fn eq(&self, other: &Self) -> bool {
-        self.unique_id == other.unique_id
-    }
-}
-
-impl Eq for GcData {}
-
-impl Ord for GcData {
-    fn cmp(&self, other: &Self) -> cmp::Ordering {
-        Ord::cmp(&self.unique_id, &other.unique_id)
-    }
-}
-
-impl PartialOrd for GcData {
-    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
-        PartialOrd::partial_cmp(&self.unique_id, &other.unique_id)
-    }
-}
-
-/// There is one `GcHandle` per `Gc<T>`. We need this metadata for collection
-#[derive(Debug)]
-pub(crate) struct GcHandle {
-    unique_id: u64,
-    /// what data is backing this handle
-    underlying_data: UnderlyingData,
-    // During what collection was this last found in a piece of GcData?
-    //     0 if this is a new piece of data
-    last_non_rooted: AtomicU64,
-}
-
-#[derive(Debug)]
-pub enum UnderlyingData {
-    Fixed(Arc<GcData>),
-    DynamicForAtomic(Arc<AtomicPtr<GcData>>),
-}
-
-impl UnderlyingData {
-    // Safe only if called with an APS Exclusive guard
-    #[inline]
-    unsafe fn with_data<F: FnOnce(&GcData)>(&self, f: F) {
-        match self {
-            UnderlyingData::Fixed(data) => f(&*data),
-            UnderlyingData::DynamicForAtomic(ptr) => {
-                let arc_ptr = ptr.load(Ordering::Relaxed);
-                f(&*arc_ptr)
-            }
-        }
-    }
-}
-
-impl Hash for GcHandle {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.unique_id.hash(state);
-    }
-}
-
-impl PartialEq for GcHandle {
-    fn eq(&self, other: &Self) -> bool {
-        self.unique_id == other.unique_id
-    }
-}
-
-impl Eq for GcHandle {}
-
-impl Ord for GcHandle {
-    fn cmp(&self, other: &Self) -> cmp::Ordering {
-        Ord::cmp(&self.unique_id, &other.unique_id)
-    }
-}
-
-impl PartialOrd for GcHandle {
-    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
-        PartialOrd::partial_cmp(&self.unique_id, &other.unique_id)
-    }
 }
 
 // TODO(issue): https://github.com/Others/shredder/issues/7
@@ -435,182 +324,9 @@ impl Collector {
         let gc_guard = self.gc_lock.lock();
         self.do_collect(gc_guard);
     }
-
-    // TODO(issue): https://github.com/Others/shredder/issues/13
-    // TODO: Remove the vectors we allocate here with an intrusive linked list
-    // TODO: Optimize memory overhead
-    #[allow(clippy::shadow_unrelated)]
-    fn do_collect(&self, gc_guard: MutexGuard<'_, ()>) {
-        // Be careful modifying this method. The tracked data and tracked handles can change underneath us
-        // Currently the state is this, as far as I can tell:
-        // - New handles are conservatively seen as roots if seen at all while we are touching handles
-        // (there is nowhere a new "secret root" can be created and then the old root stashed and seen as non-rooted)
-        // - New data is treated as a special case, and only deallocated if it existed at the start of collection
-        // - Deleted handles cannot make the graph "more connected" if the deletion was not observed
-
-        trace!("Beginning collection");
-        let _atomic_spinlock_guard = self.atomic_spinlock.lock_exclusive();
-
-        let current_collection = self
-            .tracked_data
-            .current_collection_number
-            .load(Ordering::SeqCst);
-
-        // Here we synchronize destructors: this ensures that handles in objects in the background thread are dropped
-        // Otherwise we'd see those handles as rooted and keep them around.
-        // This makes a lot of sense in the background thread (since it's totally async),
-        // but may slow direct calls to `collect`.
-        self.synchronize_destructors();
-
-        // The warrant system prevents us from scanning in-use data
-        let warrants: SegQueue<GcExclusiveWarrant> = SegQueue::new();
-
-        // eprintln!("tracked data {:?}", tracked_data);
-        // eprintln!("tracked handles {:?}", tracked_handles);
-
-        // In this step we calculate what's not rooted by marking all data definitively in a Gc
-        self.tracked_data.data.iter().par_bridge().for_each(|ele| {
-            let data = ele.key();
-
-            // If data.last_marked == 0, then it is new data. Update that we've seen this data
-            // (this step helps synchronize what data is valid to be deallocated)
-            if data.last_marked.load(Ordering::SeqCst) == 0 {
-                data.last_marked
-                    .store(current_collection - 1, Ordering::SeqCst);
-            }
-
-            if let Some(warrant) = Lockout::get_exclusive_warrant(data.clone()) {
-                // Save that warrant so things can't shift around under us
-                warrants.push(warrant);
-
-                // Now figure out what handles are not rooted
-                data.underlying_allocation.scan(|h| {
-                    h.handle_ref
-                        .last_non_rooted
-                        .store(current_collection, Ordering::SeqCst);
-                });
-            } else {
-                // eprintln!("failed to get warrant!");
-                // If we can't get the warrant, then this data must be in use, so we can mark it
-                data.last_marked.store(current_collection, Ordering::SeqCst);
-            }
-        });
-
-        // The handles that were not just marked need to be treated as roots
-        let roots = SegQueue::new();
-        self.tracked_data
-            .handles
-            .iter()
-            .par_bridge()
-            .for_each(|ele| {
-                let handle = ele.key();
-                // If the `last_non_rooted` number was not now, then it is a root
-                if handle.last_non_rooted.load(Ordering::SeqCst) != current_collection {
-                    roots.push(handle.clone());
-                }
-            });
-
-        // eprintln!("roots {:?}", roots);
-
-        // This step is dfs through the object graph (starting with the roots)
-        // We mark each object we find
-        let dfs_stack = DynQueue::new(roots);
-        dfs_stack
-            .into_par_iter()
-            .for_each(|(queue, handle)| unsafe {
-                handle.underlying_data.with_data(|data| {
-                    // If this data is new, we don't want to `Scan` it, since we may not have its Lockout
-                    // Any handles inside this could not of been seen in step 1, so they'll be rooted anyway
-                    if data.last_marked.load(Ordering::SeqCst) != 0 {
-                        // Essential note! All non-new non-warranted data is automatically marked
-                        // Thus we will never accidentally scan non-warranted data here
-                        let previous_mark =
-                            data.last_marked.swap(current_collection, Ordering::SeqCst);
-
-                        // Since we've done an atomic swap, we know we've already scanned this iff it was marked
-                        // (excluding data marked because we couldn't get its warrant, who's handles would be seen as roots)
-                        // This stops us for scanning data more than once and, crucially, concurrently scanning the same data
-                        if previous_mark != current_collection {
-                            data.last_marked.store(current_collection, Ordering::SeqCst);
-
-                            data.underlying_allocation.scan(|h| {
-                                let mut should_enque = false;
-                                h.handle_ref.underlying_data.with_data(|scanned_data| {
-                                    if scanned_data.last_marked.load(Ordering::SeqCst)
-                                        != current_collection
-                                    {
-                                        should_enque = true;
-                                    }
-                                });
-                                if should_enque {
-                                    queue.enqueue(h.handle_ref);
-                                }
-                            });
-                        }
-                    }
-                })
-            });
-        // We're done scanning things, and have established what is marked. Release the warrants
-        drop(warrants);
-
-        // Now cleanup by removing all the data that is done for
-        par_retain(&self.tracked_data.data, |data, _| {
-            // Mark the new data as in use for now
-            // This stops us deallocating data that was allocated during collection
-            if data.last_marked.load(Ordering::SeqCst) == 0 {
-                data.last_marked.store(current_collection, Ordering::SeqCst);
-            }
-
-            // If this is true, we just marked this data
-            if data.last_marked.load(Ordering::SeqCst) == current_collection {
-                // so retain it
-                true
-            } else {
-                // Otherwise we didn't mark it and it should be deallocated
-                // eprintln!("deallocating {:?}", data_ptr);
-                // Send it to the drop thread to be dropped
-                let drop_msg = DropMessage::DataToDrop(data.clone());
-                if let Err(e) = self.dropper.send_msg(drop_msg) {
-                    error!("Error sending to drop thread {}", e);
-                }
-
-                // Note: It's okay to send all the data before we've removed it from the map
-                // The destructor manages the `destructed` flag so we can never access free'd data
-
-                // Don't retain this data
-                false
-            }
-        });
-
-        // update the trigger based on the new baseline
-        self.trigger
-            .set_data_count_after_collection(self.tracked_data_count());
-
-        // update collection number
-        self.tracked_data
-            .current_collection_number
-            .fetch_add(1, Ordering::SeqCst);
-
-        drop(gc_guard);
-
-        trace!("Collection finished");
-    }
 }
 
 pub static COLLECTOR: Lazy<Arc<Collector>> = Lazy::new(Collector::new);
-
-// Helper function! Lives here because it has nowhere else to go ;-;
-fn par_retain<K, V, F: Fn(&K, &V) -> bool>(map: &DashMap<K, V>, retain_fn: F)
-where
-    K: Eq + Hash + Send + Sync,
-    V: Send + Sync,
-    F: Send + Sync,
-{
-    map.shards()
-        .iter()
-        .par_bridge()
-        .for_each(|s| s.write().retain(|k, v| retain_fn(k, v.get())));
-}
 
 #[cfg(test)]
 pub(crate) fn get_mock_handle() -> InternalGcRef {
