+++ src/concurrency/atomic_protection.rs
